pyproject.toml
[tool.poetry]
name = "regen"
version = "0.1.0"
description = ""
authors = ["nishan0000 <mdnishan006@gmail.com>"]
readme = "README.md"
packages = [{include = "regen", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "4.9.0"
numpy = "^1.26.4"
pandas = "^2.2.1"
requests = "^2.31.0"
ipykernel = "^6.29.3"
webdriver-manager = "^4.0.1"
linkedin-scraper = "^2.11.2"
anthropic = "^0.20.0"
langchain-anthropic = "^0.1.4"
langchain = "^0.1.12"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
pdfkit = "^1.0.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
